video_workflow/requirements.txt
opencv-python-headless
transformers>=4.49.0
torch
Pillow
numpy
tqdm
flask
sentencepiece
accelerate